Feat: useInput hooks를 통해 환율 반영 원화 표시
