+++ src/components/ExchangeInput.tsx
@@ -1,19 +1,8 @@
-import { ChangeEvent, useState } from 'react';
+import useInput from '../hooks/use-input';
 import { EurInfoTypes } from '../types/eur-info.type';
-import { exchangeEurToKrw } from '../utils/exchange-eur-to-krw';
-
-const regexp = /^\d*.?\d{0,2}$/;
 
 const ExchangeInput = ({ eurInfo }: { eurInfo: EurInfoTypes }) => {
-  const [eur, setEur] = useState<string>('');
-  const [krw, setKrw] = useState<string>('');
-
-  const changeHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    if (!regexp.test(e.target.value)) return;
-
-    setEur(e.target.value);
-    setKrw(exchangeEurToKrw(+e.target.value, eurInfo));
-  };
+  const { eur, krw, changeHandler } = useInput(eurInfo);
 
   return (
     <div>
