+++ src/utils/exchange-eur-to-krw.ts
@@ -1,5 +1,5 @@
-export const exchangeEurToKrw = (eur: number, eurInfo: any) => {
-  const num = eur * eurInfo.basePrice;
+export const exchangeEurToKrw = (eur: number, basePrice: number) => {
+  const num = eur * basePrice;
   return num.toLocaleString('ko-KR', {
     maximumFractionDigits: 0,
   });
